Integrate callbacks and other functionalities required for SLR #282

+++ saito-core/src/common/defs.rs
@@ -5,7 +5,7 @@ use ahash::AHashMap;
 use tokio::sync::mpsc::Sender;
 
 pub type Currency = u64;
-pub type Timestamp = u64;
+pub type Timestamp = u32;
 pub type SaitoSignature = [u8; 64];
 pub type SaitoPublicKey = [u8; 33];
 pub type SaitoPrivateKey = [u8; 32];
